requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
webdriver-manager==3.2.2
Werkzeug==1.0.1
